add tests and delete old files
