+++ apps/ex_nvr/lib/ex_nvr/pipeline/output/storage/segment_metadata.ex
@@ -1,4 +1,4 @@
-defmodule ExNVR.Pipeline.Output.Storage.Segmenter.SegmentMetadata do
+defmodule ExNVR.Pipeline.Output.Storage.SegmentMetadata do
   @moduledoc """
   A struct representing metadata about a segment
   """
